requirements.txt
streamlit==1.41.1
langchain
openai
python-dotenv
